+++ client/src/router/bootstrap.ts
@@ -1,11 +1,16 @@
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
 import type { RouteRecordRaw } from "vue-router";
 
+import { AdminSection } from "../admin/types";
 import { Logout } from "../auth/logout";
 import Invitation from "../invitation";
 
 import { router } from ".";
 
+// Admin
+const Admin = () => import("../admin/Admin.vue");
+const AdminUsers = () => import("../admin/Users.vue");
+const AdminCampaigns = () => import("../admin/Campaigns.vue");
 // Auth
 const Login = () => import("../auth/Login.vue");
 // Dashboard
@@ -98,6 +103,35 @@ const routes: RouteRecordRaw[] = [
                 path: "settings",
                 component: DashboardSettings,
                 name: "dashboard-settings",
+            },
+        ],
+    },
+    {
+        path: "/admin",
+        component: Admin,
+        meta: {
+            auth: true,
+        },
+        children: [
+            {
+                path: "",
+                redirect: { name: "admin-users" },
+            },
+            {
+                name: "admin-users",
+                path: "/admin/users",
+                component: AdminUsers,
+                meta: {
+                    adminSection: AdminSection.Users,
+                },
+            },
+            {
+                name: "admin-campaigns",
+                path: "/admin/campaigns",
+                component: AdminCampaigns,
+                meta: {
+                    adminSection: AdminSection.Campaigns,
+                },
             },
         ],
     },