+++ server/src/api/socket/__init__.py
@@ -1,5 +1,6 @@
 def load_socket_commands():
     from . import (
+        admin,  # noqa: F401
         asset,  # noqa: F401
         character,  # noqa: F401
         chat,  # noqa: F401