+++ CHANGELOG.md
@@ -10,6 +10,13 @@ tech changes will usually be stripped from release notes for the public
 
 ## Unreleased
 
+### Added
+
+-   Admin panel that can be accessed from the dashboard
+    -   Currently limited to the main admin_user as configured in the server's config
+    -   Shows roughly same content as the original separate admin client
+    -   Also shows total number of users/campaigns and a quick user create button
+
 ### Changed
 
 -   Smoothed out the scroll zoom behaviour when zoomed in furthest
@@ -18,6 +25,7 @@ tech changes will usually be stripped from release notes for the public
 ### Removed
 
 -   [server] Legacy server configs
+-   [server] Admin server - this is replaced with a new in-client admin dashboard
 
 ### Fixed